feature(devnet): Reth as sequencer (#133)
* adding reth submodule

* build op-reth docker image in init.sh

* update op-geth submodule

* add op-reth-rpc to docker-compose

* bump op-reth

* use rpc variable in docker compose

* update op-reth

* fix scripts to support op-reth-rpc

* update op-reth

* add p2p connections to op-reth-rpc

* set trusted peers in geth rpc

* update op-reth default branch

* remove op-reth submodule and update init.sh

* update readme

* feature: RPC start and stop scripts for ease of testing

* update op-geth

* improvement(test): use entrypoint in docker compose for rpc

* improvement(test): remove admin api from geth rpc

* fix: init op-geth-rpc datadir on re-start

* add new files to gitignore

* change max-outbound-peers and max-inbound-peers to 10

* feat(reth seq): add initial changes to make reth seq run

* add reth rpc flags to enable tx forwarding

* updtae gitignore

* add reth rpc flags to enable tx forwarding

* fix example.env

* update readme

* update readme

* add stop rpc script

* rm containers in kill-rpc.sh

* fix scripts

* fix scripts and configs

* fix script

* fix: add trusted peer for reth seq

+++ test/3-op-init.sh
@@ -120,10 +120,12 @@ else
     echo " ❌ state.json not found at $STATE_JSON"
 fi
 
-# init op-geth-seq and op-geth-rpc
+# init geth sequencer
+echo " 🔧 Initializing geth sequencer..."
 OP_GETH_DATADIR="$(pwd)/data/op-geth-seq"
 rm -rf "$OP_GETH_DATADIR"
 mkdir -p "$OP_GETH_DATADIR"
+
 docker compose run --no-deps --rm \
   -v "$(pwd)/$CONFIG_DIR/genesis.json:/genesis.json" \
   op-geth-seq \
@@ -138,11 +140,14 @@ docker compose run --no-deps --rm \
 echo " 🔑 Removing nodekey to generate unique node ID for other nodes..."
 rm -f "$OP_GETH_DATADIR/geth/nodekey"
 
+# init reth sequencer
+echo " 🔧 Initializing reth sequencer..."
 OP_RETH_DATADIR="$(pwd)/data/op-reth-seq"
 rm -rf "$OP_RETH_DATADIR"
 mkdir -p "$OP_RETH_DATADIR"
 INIT_LOG=$(docker compose run --no-deps --rm \
   -v "$(pwd)/$CONFIG_DIR/genesis-reth.json:/genesis.json" \
+  --entrypoint op-reth \
   op-reth-seq \
   init \
   --datadir="/datadir" \
@@ -185,13 +190,21 @@ if [ "$SEQ_TYPE" = "reth" ]; then
   echo -n "1aba031aeb5aa8aedadaf04159d20e7d58eeefb3280176c7d59040476c2ab21b" > $OP_RETH_DATADIR/discovery-secret
 fi
 
-echo "finished init op-geth-seq and op-geth-rpc"
+echo "✅ Finished init op-$SEQ_TYPE-seq and op-$RPC_TYPE-rpc."
 
 # genesis.json is too large to embed in go, so we compress it now and decompress it in go code
 gzip -c config-op/genesis.json > config-op/genesis.json.gz
 
 # Ensure prestate files exist and devnetL1.json is consistent before deploying contracts
 EXPORT_DIR="$PWD_DIR/data/cannon-data"
+SAVED_CANNON_DATA_DIR="$PWD_DIR/saved-cannon-data"
+
+if [ "$SKIP_BUILD_PRESTATE" = "true" ] && [ -d "$SAVED_CANNON_DATA_DIR" ]; then
+    echo "🔄 Skipping building op-program prestate files. Copying saved cannon data from $SAVED_CANNON_DATA_DIR to $EXPORT_DIR..."
+    cp -r $SAVED_CANNON_DATA_DIR $EXPORT_DIR
+    exit 0
+fi
+
 rm -rf $EXPORT_DIR
 mkdir -p $EXPORT_DIR
 
@@ -220,3 +233,6 @@ $DOCKER_CMD \
       /scripts/docker-install-start.sh $DOCKER_TYPE
       make -C op-program reproducible-prestate
     "
+
+echo "🔄 Copying built prestate files from $EXPORT_DIR to $SAVED_CANNON_DATA_DIR..."
+cp -r $EXPORT_DIR $SAVED_CANNON_DATA_DIR
